+++ gdscript/src/main/kotlin/config/GdConfigIcon.kt
@@ -1,0 +1,10 @@
+package config
+
+import com.intellij.openapi.util.IconLoader
+import gdscript.GdIcon
+
+object GdConfigIcon {
+
+    val FILE = IconLoader.getIcon("icons/godot_editor/FileDialog.svg", GdIcon::class.java)
+
+}